+++ aomc/aomc_test.go
@@ -1,12 +1,8 @@
 package aomc_test
 
 import (
-	"encoding/json"
 	"net/url"
-	"reflect"
 	"testing"
-
-	"github.com/philippgille/apiomat-go/aomc"
 )
 
 // FakeClient is a fake implementation of the aoms.Client interface.
@@ -26,45 +22,6 @@ var onGetVersion = func() (string, error) {
 	return "1.2.3", nil
 }
 var onGet func(string, url.Values) (string, error)
-
-// TestGetClasses tests if aomc.GetClasses leads to the correct aoms.Client call and the correct returned JSON
-func TestGetClasses(t *testing.T) {
-	// Prepare fake data and Get() implementation
-	want := []aomc.Class{
-		aomc.Class{
-			ID: "123",
-		},
-		aomc.Class{
-			ID: "456",
-		},
-	}
-	wantJsonBytes, err := json.Marshal(want)
-	stopOnError(err, t)
-	wantJson := string(wantJsonBytes)
-	moduleName := "fakeModule"
-	onGet = func(path string, params url.Values) (string, error) {
-		// Assertions
-		expectedPath := "modules/" + moduleName + "/metamodels"
-		if path != expectedPath {
-			t.Errorf("path was %q, but should be %q", path, expectedPath)
-		}
-		if params != nil {
-			t.Errorf("params was %v, but should be %v", params, nil)
-		}
-		// Assertions were okay, return fake data
-		return wantJson, nil
-	}
-	// Create fake client
-	fakeClient := FakeClient{}
-	client := aomc.NewClient(fakeClient)
-	// Call method to test
-	got, err := client.GetClasses(moduleName)
-	// Assertions
-	stopOnError(err, t)
-	if reflect.DeepEqual(got, want) == false {
-		t.Errorf("got %+v, want %+v", got, want)
-	}
-}
 
 func stopOnError(err error, t *testing.T) {
 	if err != nil {